[S08P12A809-296] Update: 팔로우 에러처리

+++ back/imfine/src/main/java/com/idle/imfine/service/user/Impl/FollowServiceImpl.java
@@ -98,6 +98,11 @@ public class FollowServiceImpl implements FollowService {
         User requester = common.getUserByUid(otherUid);
         LOGGER.info("[FollowService.allowUserRequest] 유저 정보 조회 완료");
 
+        if (followWaitRepository.existsByRequesterAndReceiver(requester, user)) {
+            LOGGER.info("[FollowService.allowUserRequest] 이미 팔로우 요청 중입니다.");
+            throw new ErrorException(FollowErrorCode.ALREADY_FOLLOW_REQUEST);
+        }
+
         Follow follow = Follow.builder()
                 .followingUser(requester)
                 .followedUser(user)
@@ -109,7 +114,6 @@ public class FollowServiceImpl implements FollowService {
         followRepository.save(follow);
         followWaitRepository.deleteByRequesterAndReceiver(requester, user);
 
-
 //        notificationService.send(user.getId(), other.getId(), 6, user.getId(), 5);
         LOGGER.info("[FollowService.allowUserRequest] 팔로우 요청을 수락했습니다.");
     }
@@ -120,8 +124,12 @@ public class FollowServiceImpl implements FollowService {
         User requester = common.getUserByUid(otherUid);
         LOGGER.info("[FollowService.declineUserRequest] 유저 정보 조회 완료");
 
-        followWaitRepository.deleteByRequesterAndReceiver(requester, user);
+        if (!followWaitRepository.existsByRequesterAndReceiver(requester, user)) {
+            LOGGER.info("[FollowService.declineUserRequest] 이미 팔로우 요청 중이 아닙니다.");
+            throw new ErrorException(FollowErrorCode.NOT_ALREADY_FOLLOW_REQUEST);
+        }
 
+        followWaitRepository.deleteByRequesterAndReceiver(requester, user);
 //        notificationService.send(user.getId(), other.getId(), 6, user.getId(), 5);
         LOGGER.info("[FollowService.declineUserRequest] 팔로우 요청을 거절했습니다.");
     }
